move storage tests to respective files
